requirements.txt
torch==2.0.0
torchvision==0.15.1
transformers==4.30.2
scikit-learn==1.3.0
Pillow==9.5.0
tqdm==4.65.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
openpyxl==3.1.2
python-docx==0.8.11
flask==2.0.1
